requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.6.1
jsonschema==4.17.3
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
PyMsgBox==1.0.9
python-dotenv==1.0.0
pytweening==1.0.7
PyYAML==6.0
typing_extensions==4.6.3
uritemplate==4.1.1
